+++ sorteia-fx/src/main/java/org/jugvale/sorteioapp/SorteaProjeto.java
@@ -1,4 +1,4 @@
-package org.jugvale;
+package org.jugvale.sorteioapp;
 /*
  * To change this license header, choose License Headers in Project Properties. To change this template file, choose Tools | Templates and open the
  * template in the editor.
