+++ sorteia-fx/src/main/java/org/jugvale/sorteioapp/spi/c4p/impl/Evento.java
@@ -1,8 +1,8 @@
-package org.jugvale.model.impl;
+package org.jugvale.sorteioapp.spi.c4p.impl;
 
 import javax.xml.bind.annotation.XmlRootElement;
 
-import org.jugvale.model.DefaultModel;
+import org.jugvale.sorteioapp.spi.c4p.model.DefaultModel;
 
 @XmlRootElement
 @SuppressWarnings("serial")
