+++ sorteia-fx/src/main/java/org/jugvale/sorteioapp/spi/c4p/impl/Nivel.java
@@ -1,4 +1,4 @@
-package org.jugvale.model.impl;
+package org.jugvale.sorteioapp.spi.c4p.impl;
 
 public enum Nivel {
 	INICIANTE, BASICO, INTERMEDIARIO, AVANCADO;
