+++ sorteia-fx/src/main/resources/logo.jpg
